Added SpawnPattern
*Spawns a basic pattern of enemies
*Made enemies not spawn after player dies
*Added some javaDoc and small fixes

+++ swipe-race-tutorial/swipe-race-tutorial/src/com/enemyShips/ScoutShip.java
@@ -11,8 +11,8 @@ import com.spacegame.MyGame;
  */
 public class ScoutShip extends EnemyShip{
 	
-	public final static float HEIGHT = 40;
-	public final static float WIDTH = 40;
+	public final static int HEIGHT = 40;
+	public final static int WIDTH = 40;
 	private final static float SHIPSPEED_X = 2.0f;
 	private final static float SHIPSPEED_Y = 3.0f;
 	private boolean movingLeft = true;
